.bashrc.alias added + install.sh changed

+++ install.sh
@@ -1,1 +1,6 @@
-ln -s nx-tools/.vimrc .vimrc
+
+if [ ! -L .vimrc ]; then 
+  ln -s nx-tools/.vimrc .vimrc
+fi
+
+grep ".bashrc.alias" .bashrc >> /dev/null || echo ". nx-tools/.bashrc.alias" >> .bashrc
